refactor(activerecord): locate validatesAssociated/validatesUniqueness next to their validators
Rails' layout keeps the `validates_associated` registrar alongside
AssociatedValidator in validations/associated.rb and
`validates_uniqueness_of` alongside UniquenessValidator in
validations/uniqueness.rb. Move our TS versions to the same files —
validations/associated.ts and validations/uniqueness.ts — and re-export
from validations.ts so the ClassMethods bundle still wires onto Base
as before. No behavior change.
